+++ Api/Data/NodeInterface.php
@@ -23,6 +23,7 @@ interface NodeInterface
     const IS_ACTIVE = 'is_active';
     const ADDITIONAL_DATA = 'additional_data';
     const SELECTED_ITEM_ID = 'selected_item_id';
+    const CUSTOMER_GROUPS = 'customer_groups';
 
     /**
      * Get node id
@@ -274,4 +275,17 @@ interface NodeInterface
      * @return $this
      */
     public function setSelectedItemId($selectedItemId);
+
+    /**
+     * Get customer groups
+     *
+     * @return int[]
+     */
+    public function getCustomerGroups();
+
+    /**
+     * @param int[] $customerGroups
+     * @return $this
+     */
+    public function setCustomerGroups($customerGroups);
 }