+++ Setup/UpgradeSchema.php
@@ -59,6 +59,10 @@ class UpgradeSchema implements UpgradeSchemaInterface
             $this->addNodeSelectedItemId($setup);
         }
 
+        if (version_compare($context->getVersion(), '0.2.8', '<')) {
+            $this->addNodeCustomerGroups($setup);
+        }
+
         $setup->endSetup();
     }
 
@@ -293,6 +297,26 @@ class UpgradeSchema implements UpgradeSchemaInterface
                 'unsigned' => true,
                 'nullable' => true,
                 'comment' => 'Selected Item Id'
+            ]
+        );
+
+        return $this;
+    }
+
+    private function addNodeCustomerGroups(SchemaSetupInterface $setup)
+    {
+        $connection = $setup->getConnection();
+        $table = $setup->getTable('snowmenu_node');
+
+        $connection->addColumn(
+            $table,
+            NodeInterface::CUSTOMER_GROUPS,
+            [
+                'type' => Table::TYPE_TEXT,
+                'length' => 255,
+                'nullable' => true,
+                'after' => NodeInterface::SELECTED_ITEM_ID,
+                'comment' => "Customer Groups Serialized"
             ]
         );